backend/requirements.txt
langchain-huggingface
langchain-openai
langchain-community
langchain-core
langgraph
pydantic
fastapi>=0.100.0
python-multipart>=0.0.6
uvicorn>=0.23.0
pydantic>=2.0.0
typing-extensions>=4.7.1
starlette>=0.27.0
uvicorn>=0.13.0
dotenv
pymongo>=4.5.0
motor>=3.3.1
# Legacy database support
psycopg2-binary>=2.9.5
# For the scraper
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
cachetools>=5.2.0
copilotkit
